refactoring singular file into directory due to growing code base
